Rename Seaweed to seaweed
