+++ src/lib/index.ts
@@ -1,7 +1,7 @@
 // noinspection JSUnusedGlobalSymbols
 
 export { default as PineappleLayoutBase } from "./components/pineapple/PineappleBaseLayout.svelte";
-export { default as SeaweedTemplate } from "./template/Seaweed/SeaweedTemplate.svelte";
+export { default as SeaweedTemplate } from "$pkg/template/seaweed/SeaweedTemplate.svelte";
 export { default as LazyAsset } from "./components/LazyAsset.svelte";
 export { default as Card } from "./components/Card.svelte";
 export { default as SocialSection } from "./components/SocialSection.svelte";
